pyproject.toml
[project]
name = "forum-memory-agent"
version = "1.0.0"
description = "Knowledge forum with pluggable memory agent system"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.27",
    "sqlmodel>=0.0.22",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "pydantic>=2.6",
    "pydantic-settings>=2.1",
    "openai>=1.12",
    "elasticsearch>=8.9,<9.0",
    "requests>=2.31",
    "python-multipart>=0.0.9",
    "dagster>=1.6",
    "dagster-webserver>=1.6",
    "slowapi>=0.1.9",
    "pyjwt>=2.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
    "ruff>=0.3",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["forum_memory*"]

[tool.ruff]
line-length = 120
target-version = "py311"
